+++ packages/boilerplate/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @calatrava/boilerplate
 
+## 0.0.10
+
+### Patch Changes
+
+- Use custom delimiter in all packages
+
 ## 0.0.9
 
 ### Patch Changes